# packages/contracts/openapi/plotify-chat.v1.json
{
  "components": {
    "schemas": {
...
      "BotResponse": {
        "properties": {
          "bot_username": {
            "title": "Bot Username",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          }
        },
        "required": [
          "bot_username",
          "is_active"
        ],
        "title": "BotResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegisterBotRequest": {
        "properties": {
          "bot_token": {
            "title": "Bot Token",
            "type": "string"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          }
        },
        "required": [
          "bot_token",
          "organization_id"
        ],
        "title": "RegisterBotRequest",
        "type": "object"
      },
      "ReservationPayload": {
        "description": "Datos del cliente y la operación, almacenados como JSONB.",
        "properties": {
          "cliente_nombre": {
            "minLength": 2,
            "title": "Cliente Nombre",
            "type": "string"
          },
          "cliente_run": {
            "minLength": 7,
            "title": "Cliente Run",
            "type": "string"
          },
          "fecha_firma": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Firma"
          },
          "notaria": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notaria"
          },
          "valor_reserva": {
            "exclusiveMinimum": 0.0,
            "title": "Valor Reserva",
            "type": "number"
          }
        },
        "required": [
          "cliente_nombre",
          "cliente_run",
          "valor_reserva"
        ],
        "title": "ReservationPayload",
        "type": "object"
      },
      "ReservationRequest": {
        "description": "Payload completo que envía el Frontend para solicitar una reserva.",
        "properties": {
          "lot_id": {
            "title": "Lot Id",
            "type": "string"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/ReservationPayload"
          },
          "vendor_id": {
            "title": "Vendor Id",
            "type": "string"
          },
          "vendor_name": {
            "title": "Vendor Name",
            "type": "string"
          },
          "vendor_phone": {
            "title": "Vendor Phone",
            "type": "string"
          },
          "vendor_platform": {
            "pattern": "^(telegram|whatsapp)$",
            "title": "Vendor Platform",
            "type": "string"
          }
        },
        "required": [
          "lot_id",
          "organization_id",
          "vendor_id",
          "vendor_name",
          "vendor_phone",
          "vendor_platform",
          "payload"
        ],
        "title": "ReservationRequest",
        "type": "object"
      },
      "ReservationResponse": {
        "description": "Respuesta del endpoint al Frontend.",
        "properties": {
          "approval_id": {
            "title": "Approval Id",
            "type": "string"
          },
          "message": {
            "default": "Solicitud enviada al administrador.",
            "title": "Message",
            "type": "string"
          },
          "status": {
            "default": "pending",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "approval_id"
        ],
        "title": "ReservationResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Internal-Secret",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Plotify Messaging Engine",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/approvals/request-reservation": {
      "post": {
        "description": "Recibe una solicitud de reserva desde el Frontend.\n1. Valida que el lote esté disponible.\n2. Valida que no haya otra solicitud pendiente para ese lote.\n3. Inserta en approval_requests.\n4. Encola notificación al Admin vía Redis.",
        "operationId": "requestReservationApproval",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Request Reservation",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/v1/bots/register": {
      "post": {
        "operationId": "register_bot_api_v1_bots_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterBotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Bot",
        "tags": [
          "bots"
        ]
      }
    },
    "/api/v1/bots/{org_id}": {
      "delete": {
        "operationId": "delete_bot_api_v1_bots__org_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Bot",
        "tags": [
          "bots"
        ]
      },
      "get": {
        "operationId": "get_bot_api_v1_bots__org_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/BotResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Bot Api V1 Bots  Org Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Bot",
        "tags": [
          "bots"
        ]
      }
    },
...
  }
}